moved constants from generation to root
