+++ maze_dataset/constants.py
