+++ maze_dataset/plotting/plot_maze.py
@@ -11,7 +11,7 @@ from matplotlib.cm import ScalarMappable
 from matplotlib.colors import ListedColormap, Normalize
 from muutils.tensor_utils import NDArray
 
-from maze_dataset.generation.constants import Coord, CoordArray, CoordList
+from maze_dataset.constants import Coord, CoordArray, CoordList
 from maze_dataset.generation.lattice_maze import (
     Coord,
     CoordArray,
