+++ tests/unit/maze_dataset/generation/test_maze_dataset.py
@@ -3,6 +3,7 @@ import copy
 import numpy as np
 import pytest
 
+from maze_dataset.constants import CoordArray
 from maze_dataset.dataset.dataset import (
     register_dataset_filter,
     register_filter_namespace_for_dataset,
@@ -12,7 +13,6 @@ from maze_dataset.dataset.maze_dataset import (
     MazeDatasetConfig,
     register_maze_filter,
 )
-from maze_dataset.generation.constants import CoordArray
 from maze_dataset.generation.lattice_maze import SolvedMaze
 from maze_dataset.utils import bool_array_from_string
